requirements.txt
flask
python-dotenv
requests
werkzeug